backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.25.0

